+++ org.mixedrealitytoolkit.uxcomponents/CHANGELOG.md
@@ -8,3 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 * Fixed support for UPM package publishing in the Unity Asset Store.
 * Fixed Bug 643 - Experimental SimpleEmptyButton and SimpleActionButton prefabs have missing "See It Say It Label" GameObject reference
+
+### Added
+
+* Added touch support for the NonNativeKeyboard.