+++ org.mixedrealitytoolkit.uxcomponents/CHANGELOG.md
@@ -2,9 +2,17 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
-## [3.2.0-development] - 2024-02-04
+## [3.2.0-development] - 2024-03-20
+
+### Added
+
+* Added Empty, Action, Action with Checkbox, and ToggleSwitch Experimental buttons prefabs (#649)
+* Added corresponding MenuItems for use in Editor Tools (#649)
+* Added CanvasUITearsheetDynamicFrontplate scene (#649)
+* Added SimpleButton + demo scene (#635)
 
 ### Fixed
 
-* Fixed support for UPM package publishing in the Unity Asset Store.
-* Fixed Bug 643 - Experimental SimpleEmptyButton and SimpleActionButton prefabs have missing "See It Say It Label" GameObject reference
+* Fixed support for UPM package publishing in the Unity Asset Store. (#519)
+* Fixed Bug #643 - Experimental SimpleEmptyButton and SimpleActionButton prefabs have missing "See It Say It Label" GameObject reference (#646)
+* Auto disable button's colliders when `StatefulInteractable` is disabled (#626)